Integrate feedback from @johnthiriet
