+++ MvvmHelpers/Exceptions/InvalidCommandParameterException.cs
@@ -10,10 +10,10 @@ namespace MvvmHelpers.Exceptions
 		/// <summary>
 		/// Initializes a new instance of the <see cref="T:MvvmHelpersInvalidCommandParameterException"/> class.
 		/// </summary>
-		/// <param name="excpectedType">Excpected parameter type for AsyncCommand.Execute.</param>
+		/// <param name="expectedType">Expected parameter type for AsyncCommand.Execute.</param>
 		/// <param name="actualType">Actual parameter type for AsyncCommand.Execute.</param>
 		/// <param name="innerException">Inner Exception</param>
-		public InvalidCommandParameterException(Type excpectedType, Type actualType, Exception innerException) : base(CreateErrorMessage(excpectedType, actualType), innerException)
+		public InvalidCommandParameterException(Type expectedType, Type actualType, Exception innerException) : base(CreateErrorMessage(expectedType, actualType), innerException)
 		{
 
 		}
@@ -21,14 +21,14 @@ namespace MvvmHelpers.Exceptions
 		/// <summary>
 		/// Initializes a new instance of the <see cref="T:MvvmHelpers.InvalidCommandParameterException"/> class.
 		/// </summary>
-		/// <param name="excpectedType">Excpected parameter type for AsyncCommand.Execute.</param>
+		/// <param name="expectedType">Expected parameter type for AsyncCommand.Execute.</param>
 		/// <param name="actualType">Actual parameter type for AsyncCommand.Execute.</param>
-		public InvalidCommandParameterException(Type excpectedType, Type actualType) : base(CreateErrorMessage(excpectedType, actualType))
+		public InvalidCommandParameterException(Type expectedType, Type actualType) : base(CreateErrorMessage(expectedType, actualType))
 		{
 
 		}
 
-		static string CreateErrorMessage(Type excpectedType, Type actualType) =>
-			$"Invalid type for parameter. Expected Type {excpectedType}, but received Type {actualType}";
+		static string CreateErrorMessage(Type expectedType, Type actualType) =>
+			$"Invalid type for parameter. Expected Type: {expectedType}, but received Type: {actualType}";
 	}
 }
